% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_brecha.R
\name{calc_brecha}
\alias{calc_brecha}
\title{Calcular brecha porcentual}
\usage{
calc_brecha(valor1, valor2)
}
\arguments{
\item{valor1}{El valor del grupo de interés (ej. Mujeres).}

\item{valor2}{El valor de referencia o base (ej. Varones).}
}
\value{
Un valor numérico que representa la brecha porcentual.
}
\description{
Esta función calcula la diferencia porcentual entre dos valores.
Es la fórmula estándar de la DES para medir brechas.
}
\examples{
calc_brecha(80, 100)
}
