% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dir_to_qs2.R
\name{convert_dir_to_qs2}
\alias{convert_dir_to_qs2}
\title{Convertir un directorio entero a formato qs2}
\usage{
convert_dir_to_qs2(path = ".", recursive = FALSE)
}
\arguments{
\item{path}{Ruta a la carpeta que contiene los archivos.}

\item{recursive}{Logico. ¿Debe buscar tambien en subcarpetas?}
}
\description{
Busca todos los archivos .csv y .qs en una carpeta y los convierte a .qs2.
}
