% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_qs2.R
\name{convert_to_qs2}
\alias{convert_to_qs2}
\title{Convertir a formato qs2}
\usage{
convert_to_qs2(x, output_file = NULL)
}
\arguments{
\item{x}{Un objeto (data.frame) o una ruta (string) a un archivo .csv o .qs.}

\item{output_file}{Ruta de destino. Si es NULL, se genera automaticamente para archivos.}
}
\description{
Lee un archivo (.csv o .qs) o recibe un data.frame y lo guarda como .qs2
usando el motor de qs2 para maxima eficiencia.
}
